Remove HTTP_HOST header
This header causes problems in development environments and could cause
problems in production environments were our configuration to change.

The rack proxy passes the HTTP_HOST header by default into the proxied
request. This will mean that a request to say
authenticating-proxy.dev.gov.uk will be proxied to draft-origin.dev.gov.uk
with a HTTP_HOST header of authenticating-proxy.dev.gov.uk. This would
mean that a virtual host would interpret the request as being intended
for the authenticating-proxy.dev.gov.uk host and direct it there. This
causes an eternal loop.

In production this is likely not a problem because we upstream to router
at localhost:3054. Requests to this likely pay no attention to the
HTTP_HOST header.

I chose to delete the header rather than set it to the upstream_uri host
as was done in [1] as setting it to a different value can alter the
destination of signon redirects.

[1]: https://github.com/alphagov/authenticating-proxy/commit/14304161f0fb48d30a9fe829abf9c1e704b64ebd